DEV: create seperate module for SegmentImageTab
